fix(auto_source): close assure LOC wiring and recall locks
Shrink the heuristic surface below origin/master by relocating
classifiers/extractor helpers, delete the Discovery husk, wire Html
through Scoring::Engine, type Observation, and restore intentional
recall with a score floor plus Engine specs.
