+++ lib/html2rss/link_destination.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Html2rss
+  ##
+  # Destination URL / anchor-text classification shared by Scoring.
+  module LinkDestination
+  end
+end
