+++ lib/html2rss/scoring/engine.rb
@@ -5,31 +5,35 @@ module Html2rss
     ##
     # Ranks AutoSource::Segment instances with an explicit feature registry (no send).
     class Engine
-      # Maximum segments to extract after ranking (lazy extraction guard).
-      TOP_K = 50
+      # Soft cap after score-floor filtering (lazy extraction guard; breadth > blind top-50).
+      TOP_K = 99
+      # Minimum composite score retained by {#rank_top} (precision floor).
+      SCORE_FLOOR = 0.0
 
       # Quality feature triples: [FeatureId, predicate, weight].
       QUALITY_RULES = [
-        [:title_word_count_ge3, ->(o) { o[:title_word_count] >= 3 }, 40],
-        [:title_word_count_ge7, ->(o) { o[:title_word_count] >= 7 }, 15],
-        [:path_length_gt6, ->(o) { o[:path_length] > 6 }, 20],
-        [:content_path, ->(o) { o[:content_path] }, 15],
-        [:publish_marker, ->(o) { o[:publish_marker] }, 15],
-        [:descriptive_context, ->(o) { o[:descriptive_context] }, 10],
-        [:article_container, ->(o) { o[:article_container] }, 10],
-        [:content_tokens, ->(o) { o[:content_tokens] }, 10]
+        [:title_word_count_ge3, ->(o) { o.title_word_count >= 3 }, 40],
+        [:title_word_count_ge7, ->(o) { o.title_word_count >= 7 }, 15],
+        [:path_length_gt6, ->(o) { o.path_length > 6 }, 20],
+        [:content_path, lambda(&:content_path), 15],
+        [:publish_marker, lambda(&:publish_marker), 15],
+        [:descriptive_context, lambda(&:descriptive_context), 10],
+        [:article_container, lambda(&:article_container), 10],
+        [:content_tokens, lambda(&:content_tokens), 10]
       ].freeze
 
       # Junk feature triples: [FeatureId, predicate, weight].
       JUNK_RULES = [
-        [:non_content_utility_path, ->(o) { o[:utility_path] && !o[:content_path] && !o[:strong_post_suffix] }, 25],
-        [:utility_prefix_title_short, ->(o) { o[:utility_prefix_title] && o[:title_word_count] <= 6 }, 15],
-        [:shallow, ->(o) { o[:shallow] }, 10],
-        [:weak_container, ->(o) { !o[:publish_marker] && !o[:descriptive_context] }, 10],
-        [:recommended_title_non_content, ->(o) { o[:recommended_title] && !o[:content_path] }, 10],
-        [:high_confidence_junk_path, ->(o) { o[:high_confidence_junk_path] }, 5],
-        [:junk_tokens, ->(o) { o[:junk_tokens] }, 15]
+        [:non_content_utility_path, ->(o) { o.utility_path && !o.content_path && !o.strong_post_suffix }, 25],
+        [:utility_prefix_title_short, ->(o) { o.utility_prefix_title && o.title_word_count <= 6 }, 15],
+        [:shallow, lambda(&:shallow), 10],
+        [:weak_container, ->(o) { !o.publish_marker && !o.descriptive_context }, 10],
+        [:recommended_title_non_content, ->(o) { o.recommended_title && !o.content_path }, 10],
+        [:high_confidence_junk_path, lambda(&:high_confidence_junk_path), 5],
+        [:junk_tokens, lambda(&:junk_tokens), 15]
       ].freeze
+
+      (QUALITY_RULES + JUNK_RULES).each { |feature_id, _, _| FeatureId.assert!(feature_id) }
 
       # @param link_resolver [LinkResolver]
       def initialize(link_resolver:)
@@ -46,13 +50,13 @@ module Html2rss
       end
 
       ##
-      # Filters hard-junk but preserves discovery order (used by list/cluster fallback scrapers).
+      # Filters hard-junk but preserves discovery order (list/cluster scrapers).
       #
       # @param segments [Array<Html2rss::AutoSource::Segment>]
       # @param limit [Integer]
       # @return [Array<RankedSegment>]
       def select_eligible(segments, limit: TOP_K)
-        segments.filter_map { |segment| rank_one(segment) }
+        segments.filter_map { |segment| rank_one(segment, hard_junk_mode: :lenient) }
                 .sort_by(&:position)
                 .first(limit)
       end
@@ -60,19 +64,17 @@ module Html2rss
       ##
       # @param segments [Array<Html2rss::AutoSource::Segment>]
       # @param limit [Integer]
-      # @return [Array<RankedSegment>]
+      # @return [Array<RankedSegment>] score-floor filtered, then capped
       def rank_top(segments, limit: TOP_K)
-        rank(segments).first(limit)
+        rank(segments).select { |ranked| ranked.final_score >= SCORE_FLOOR }.first(limit)
       end
 
       private
 
-      def rank_one(segment) # rubocop:disable Metrics/MethodLength
+      def rank_one(segment, hard_junk_mode: :strict) # rubocop:disable Metrics/MethodLength
         facts = segment.primary_link ? @link_resolver.destination_facts(segment.primary_link) : nil
-        return if facts&.high_confidence_junk_path && segment.primary_link
-
         obs = @assessor.call(segment.root_node, segment.primary_link, destination_facts: facts)
-        return if hard_junk?(obs)
+        return if obs.hard_junk?(mode: hard_junk_mode)
 
         quality, quality_parts = apply_rules(QUALITY_RULES, obs)
         junk, junk_parts = apply_rules(JUNK_RULES, obs)
@@ -93,16 +95,6 @@ module Html2rss
           parts[feature_id] = weight if predicate.call(obs)
         end
         [parts.values.sum, parts]
-      end
-
-      def hard_junk?(obs) # rubocop:disable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
-        weak = [obs[:article_container], obs[:publish_marker], obs[:descriptive_context], obs[:content_path]]
-               .count(&:itself) < 2
-
-        obs[:high_confidence_junk_path] ||
-          (obs[:selected_anchor_present] && obs[:recommended_title] && obs[:shallow] && weak) ||
-          (obs[:selected_anchor_present] && obs[:utility_prefix_title] &&
-            obs[:high_confidence_utility_destination] && weak)
       end
     end
   end
